requirements.txt
APScheduler==3.11.1
asgiref==3.11.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
crispy-tailwind==1.0.3
cryptography==46.0.3
Django==5.2.8
django-crispy-forms==2.5
django-phonenumber-field==8.4.0
django-widget-tweaks==1.5.0
exceptiongroup==1.2.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
packaging==24.2
phonenumbers==9.0.19
pillow==11.0.0
playwright==1.51.0
pycparser==2.22
pyee==12.1.1
PySocks==1.7.1
pytesseract==0.3.13
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
tenacity==9.1.2
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
